+++ Assets/Room02_Operating/Clues/ClueAdminGuideOverlay.cs
@@ -100,6 +100,7 @@ namespace EscapeRoom
                 }
 
                 storyManager?.GrantEscapeKeyFromAdminSkip();
+                EscapeChaseQTE.StartOrCreate();
             }
             finally
             {